refactor: user_modelをフォルダに移動
