+++ lib/domain/models/user_model/user_model.dart
